wip Fix postgres parameters
These were never getting set (based on cert/password toggle) because we
never passed in options to PeConf.

Also switching to default to cert form, since we hardly ever want to set
passwords.

THIS IS INCOMPLETE!

Need to fix the passwords to be on the profile::database
and to ensure that cert_auth is false...